requirements.txt
langchain
gradio
python-dotenv
openai